Revert "chore: rewrite role resolver (#16373)" (#16431)

+++ airbyte-server/src/main/kotlin/io/airbyte/server/apis/publicapi/controllers/TagsController.kt
@@ -4,10 +4,10 @@
 
 package io.airbyte.server.apis.publicapi.controllers
 
-import io.airbyte.commons.auth.AuthRoleConstants
-import io.airbyte.commons.server.authorization.RoleResolver
+import io.airbyte.api.model.generated.PermissionType
+import io.airbyte.commons.server.authorization.ApiAuthorizationHelper
+import io.airbyte.commons.server.authorization.Scope
 import io.airbyte.commons.server.scheduling.AirbyteTaskExecutors
-import io.airbyte.commons.server.support.AuthenticationId
 import io.airbyte.commons.server.support.CurrentUserService
 import io.airbyte.publicApi.server.generated.apis.PublicTagsApi
 import io.airbyte.publicApi.server.generated.models.TagCreateRequest
@@ -28,17 +28,24 @@ import java.util.UUID
 @Controller(API_PATH)
 @Secured(SecurityRule.IS_AUTHENTICATED)
 open class TagsController(
-  private val roleResolver: RoleResolver,
+  private val apiAuthorizationHelper: ApiAuthorizationHelper,
   private val currentUserService: CurrentUserService,
   private val tagService: TagService,
   private val trackingHelper: TrackingHelper,
 ) : PublicTagsApi {
-  @Secured(AuthRoleConstants.WORKSPACE_EDITOR)
   override fun publicCreateTag(tagCreateRequest: TagCreateRequest): Response {
+    val userId: UUID = currentUserService.currentUser.userId
+    apiAuthorizationHelper.checkWorkspacePermission(
+      tagCreateRequest.workspaceId.toString(),
+      Scope.WORKSPACE,
+      userId,
+      PermissionType.WORKSPACE_EDITOR,
+    )
+
     val tag =
       trackingHelper.callWithTracker({
         tagService.createTag(tagCreateRequest)
-      }, TAGS_PATH, POST, currentUserService.currentUser.userId)
+      }, TAGS_PATH, POST, userId)
 
     return Response
       .status(Response.Status.OK.statusCode)
@@ -49,16 +56,17 @@ open class TagsController(
   override fun publicDeleteTag(tagId: UUID): Response {
     // We need to manually fetch the workspaceId here because there is no Tag scope in the ApiAuthorizationHelper
     val workspaceId: UUID = tagService.getTag(tagId).workspaceId
-
-    roleResolver
-      .Request()
-      .withCurrentUser()
-      .withRef(AuthenticationId.WORKSPACE_ID, workspaceId)
-      .requireRole(AuthRoleConstants.WORKSPACE_EDITOR)
+    val userId: UUID = currentUserService.currentUser.userId
+    apiAuthorizationHelper.checkWorkspacePermission(
+      workspaceId.toString(),
+      Scope.WORKSPACE,
+      userId,
+      PermissionType.WORKSPACE_EDITOR,
+    )
 
     trackingHelper.callWithTracker({
       tagService.deleteTag(tagId, workspaceId)
-    }, TAGS_PATH, GET, currentUserService.currentUser.userId)
+    }, TAGS_PATH, GET, userId)
 
     return Response
       .status(Response.Status.NO_CONTENT)
@@ -68,17 +76,18 @@ open class TagsController(
   override fun publicGetTag(tagId: UUID): Response {
     // We need to manually fetch the workspaceId here because there is no Tag scope in the ApiAuthorizationHelper
     val workspaceId: UUID = tagService.getTag(tagId).workspaceId
-
-    roleResolver
-      .Request()
-      .withCurrentUser()
-      .withRef(AuthenticationId.WORKSPACE_ID, workspaceId)
-      .requireRole(AuthRoleConstants.WORKSPACE_READER)
+    val userId: UUID = currentUserService.currentUser.userId
+    apiAuthorizationHelper.checkWorkspacePermission(
+      workspaceId.toString(),
+      Scope.WORKSPACE,
+      userId,
+      PermissionType.WORKSPACE_READER,
+    )
 
     val tag =
       trackingHelper.callWithTracker({
         tagService.getTag(tagId)
-      }, TAGS_PATH, GET, currentUserService.currentUser.userId)
+      }, TAGS_PATH, GET, userId)
 
     return Response
       .status(Response.Status.OK.statusCode)
@@ -88,20 +97,17 @@ open class TagsController(
 
   @ExecuteOn(AirbyteTaskExecutors.PUBLIC_API)
   override fun publicListTags(workspaceIds: List<UUID>?): Response {
-    // If workspace IDs were given, then verify the user has access to those workspaces.
-    // If none were given, then the TagService will determine the workspaces for the current user.
-    if (!workspaceIds.isNullOrEmpty()) {
-      roleResolver
-        .Request()
-        .withCurrentUser()
-        .withWorkspaces(workspaceIds)
-        .requireRole(AuthRoleConstants.WORKSPACE_READER)
-    }
-
+    val userId: UUID = currentUserService.currentUser.userId
+    apiAuthorizationHelper.checkWorkspacesPermission(
+      workspaceIds?.let { it.map { it.toString() } } ?: emptyList(),
+      Scope.WORKSPACES,
+      userId,
+      PermissionType.WORKSPACE_READER,
+    )
     val tags =
       trackingHelper.callWithTracker({
         tagService.listTags(workspaceIds ?: emptyList())
-      }, TAGS_PATH, GET, currentUserService.currentUser.userId)
+      }, TAGS_PATH, GET, userId)
 
     return Response
       .status(Response.Status.OK.statusCode)
@@ -116,16 +122,18 @@ open class TagsController(
     // We need to manually fetch the workspaceId here because there is no Tag scope in the ApiAuthorizationHelper
     val workspaceId: UUID = tagService.getTag(tagId).workspaceId
 
-    roleResolver
-      .Request()
-      .withCurrentUser()
-      .withRef(AuthenticationId.WORKSPACE_ID, workspaceId)
-      .requireRole(AuthRoleConstants.WORKSPACE_EDITOR)
+    val userId: UUID = currentUserService.currentUser.userId
+    apiAuthorizationHelper.checkWorkspacePermission(
+      workspaceId.toString(),
+      Scope.WORKSPACE,
+      userId,
+      PermissionType.WORKSPACE_EDITOR,
+    )
 
     val tag =
       trackingHelper.callWithTracker({
         tagService.updateTag(tagId, workspaceId, tagPatchRequest)
-      }, TAGS_PATH, POST, currentUserService.currentUser.userId)
+      }, TAGS_PATH, POST, userId)
 
     return Response
       .status(Response.Status.OK.statusCode)
